refactor(tui): separate input and rendering responsibilities
